requirements.txt
pillow
pytesseract
numpy
requests
opencv-python
easyocr
gspread
oauth2client
torch
torchvision